AI_AGENT/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
click>=8.0
rich>=10.0
